[#noissue] Fix static analysis issues

+++ web/src/main/java/com/navercorp/pinpoint/web/vo/LoadFactor.java
@@ -35,7 +35,7 @@ public class LoadFactor {
     private final Logger logger = LoggerFactory.getLogger(this.getClass());
 
     public static final Integer SLOT_VERY_SLOW = Integer.MAX_VALUE - 1;
-    public static final Integer SLOT_ERROR = (int)Integer.MAX_VALUE;
+    public static final Integer SLOT_ERROR = Integer.MAX_VALUE;
 
 //    /**
 //     * <pre>
@@ -139,7 +139,7 @@ public class LoadFactor {
 //        long value = histogramSummary.containsKey(responseTimeslot) ? histogramSummary.get(responseTimeslot) + callCount : callCount;
 //        histogramSummary.put(responseTimeslot, value);
 
-        /**
+        /*
          * <pre>
          * timeseriesValueList :
          * list[response_slot_no + 0] = value<timestamp, call count>
